requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.2
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
distlib==0.3.1
Django==3.1.2
django-cors-headers==3.5.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
filelock==3.0.12
idna==2.10
inflection==0.5.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.3
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jupyter-client==6.1.7
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==20.7
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytz==2020.4
pyzmq==19.0.2
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
selenium==3.141.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.0.31
wcwidth==0.2.5
wrapt==1.12.1
